Fix wave: advertise outline layer, refresh it on recheck, plug silent gaps
- skills/run/SKILL.md: add outline to frontmatter description and the
  subcommand table; recheck now archives-then-refreshes the outline and
  hands outline.json to the developmental editor on the targeted and
  restructured verdicts, mirroring assess; swept cross-references.
- README.md: outline is now a fourth layer in the intro, command list,
  flowchart, and the agents/ file listing (fiction-outliner.md and
  fiction-continuity-joiner.md were both missing).
- go/internal/cli/outline.go: a published_through boundary that matches
  no rendered section now appends a note instead of silently vanishing;
  new test covers it.
- go/internal/schemas/outline_schema.go: order validation rejects any
  non-integral value instead of truncating; new test covers order 1.5.
- go/internal/schemas/domain_loader.go: DraftStageNames delegates to
  namedFields instead of duplicating it.
- go/internal/cli/state.go, dispatch.go: consistency fix and a usage-line
  omission (archive) fixed.
- skills/new-domain/SKILL.md: note that a serial-shaped custom domain may
  want a published boundary too, pointing at serial-fiction's approach.

+++ go/internal/schemas/outline_schema.go
@@ -120,7 +120,7 @@ func ValidateOutlineSection(reportRaw interface{}, rowFields, sectionFields []st
 		// Order is 1-based and sequential. It is the row's only identity
 		// within a section, so a gap or a repeat makes the join ambiguous.
 		order, isNumber := scene["order"].(float64)
-		if !isNumber || int(order) != i+1 {
+		if !isNumber || order != float64(i+1) {
 			errors = append(errors, fmt.Sprintf("%s: order %s must be %d (1-based, sequential, matching position)", prefix, pyRepr(scene["order"]), i+1))
 		}
